Forgot to remove this line.
Removed redundant parentheses.

Changed ticket price from float to decimal

Removed redundant login_required before_filter.

+++ ticketing/db/migrate/20100611161935_create_trips.rb
@@ -7,7 +7,7 @@ class CreateTrips < ActiveRecord::Migration
       t.time :departure
       t.time :arrival
       t.time :return
-      t.float :ticket_price
+      t.decimal :ticket_price
       t.integer :sales_lead
       t.text :comment
       t.integer :return_trip
